Factor out hotdocs_cmp.py and document it.
